+++ src/ska_low_mccs_spshw/tile/tile_device.py
@@ -1295,6 +1295,21 @@ class MccsTile(MccsBaseDevice[TileComponentManager]):
         """
         return self._attribute_state["clock_managers"].read()
 
+    @attribute(
+        dtype="DevLong",
+        label="ddr_write_size",
+    )
+    def ddr_write_size(self: MccsTile) -> int:
+        """
+        Return the ddr write size obtained from running start_antenna_buffer.
+
+        :example:
+            >>> tile.ddr_write_size
+
+        :return: ddr write size of a frame
+        """
+        return self.component_manager.ddr_write_size
+
     # @attribute(
     #     dtype="DevString",
     #     label="ddr_rd_cnt",